Fix cmakeFiles. Fix code so that it compiles with gcc and clang

+++ src/CartesianProduct.cpp
@@ -43,13 +43,13 @@ namespace pgs
   Segment CartesianProduct::getValue(Eigen::Ref<Eigen::VectorXd> val, size_t i) const
   {
     assert(i < submanifolds_.size() && "invalid index");
-    return val.segment(startIndexR_[i],submanifolds_[i]->representationDim());
+    return val.segment(static_cast<long> (startIndexR_[i]),static_cast<long> (submanifolds_[i]->representationDim()));
   }
 
   ConstSegment CartesianProduct::getValueConst(const Eigen::Ref<const Eigen::VectorXd>& val, size_t i) const
   {
     assert(i < submanifolds_.size() && "invalid index");
-    return val.segment(startIndexR_[i], submanifolds_[i]->representationDim());
+    return val.segment(static_cast<long> (startIndexR_[i]), static_cast<long> (submanifolds_[i]->representationDim()));
   }
 
   void CartesianProduct::plus_(Eigen::Ref<Eigen::VectorXd> out, const Eigen::Ref<const Eigen::VectorXd>& x, const Eigen::Ref<const Eigen::VectorXd>& v) const
@@ -58,7 +58,7 @@ namespace pgs
     {
       submanifolds_[i]->plus(getValue(out, i), 
                               getValueConst(x, i), 
-                              v.segment(startIndexT_[i], submanifolds_[i]->dim()));
+                              v.segment(static_cast<long> (startIndexT_[i]),static_cast<long> ( submanifolds_[i]->dim())));
     }
   }
